pyproject.toml
[project]
name = "scalablerag-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=5.5.0",
    "bert-score>=0.3.13",
    "datasets>=4.0.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "ipykernel>=6.30.1",
    "jinja2>=3.1.6",
    "langchain>=0.3.27",
    "langchain-cohere>=0.4.6",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-pinecone>=0.2.12",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pinecone-client>=6.0.0",
    "plotly>=6.3.0",
    "pyarrow>=21.0.0",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "rouge-score>=0.1.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "tiktoken>=0.11.0",
    "unstructured>=0.18.14",
    "uvicorn>=0.35.0",
    "wandb>=0.21.4",
]
